Add document batching method to DocumentCorpus class & basic test for it
