+++ annif/corpus/__init__.py
@@ -3,7 +3,6 @@
 
 from .combine import CombinedCorpus
 from .document import (
-    BatchingDocumentCorpus,
     DocumentDirectory,
     DocumentFile,
     DocumentList,
@@ -28,5 +27,4 @@ __all__ = [
     "CombinedCorpus",
     "TransformingDocumentCorpus",
     "LimitingDocumentCorpus",
-    "BatchingDocumentCorpus",
 ]
