+++ annif/corpus/types.py
@@ -2,6 +2,7 @@
 
 import abc
 import collections
+from itertools import islice
 
 Document = collections.namedtuple("Document", "text subject_set")
 
@@ -14,6 +15,16 @@ class DocumentCorpus(metaclass=abc.ABCMeta):
     def documents(self):
         """Iterate through the document corpus, yielding Document objects."""
         pass  # pragma: no cover
+
+    def doc_batches(self, batch_size):
+        """Iterate through the document corpus in batches of a given size, yielding
+        lists of Document objects."""
+        it = iter(self.documents)
+        while True:
+            docs_batch = list(islice(it, batch_size))
+            if not docs_batch:
+                return
+            yield docs_batch
 
     def is_empty(self):
         """Check if there are no documents to iterate."""
